Make model group
make model group and independent swift file to do MVC
